fail_coro: split common part into a separate file
